backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.8.1
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
git-filter-repo==2.47.0
idna==3.10
iniconfig==2.1.0
oauthlib==3.2.2
packaging==25.0
phonenumbers==9.0.5
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
